pyproject.toml
[tool.poetry]
name = "tensor"
version = "0.1.0"
description = ""
authors = ["Denis Butylin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "7.2.0"
allure-pytest = "2.12.0"
requests = "^2.28.2"
selenium = "4.9.0"
types-requests = "^2.31"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
black = "^23.3.0"
commitizen = "^2.42.1"
cyclonedx-bom = "^3.11.0"
flake8 = "^6.0.0"
flake8-annotations = "^3.0.0"
flake8-annotations-complexity = "^0.0.7"
flake8-annotations-coverage = "^0.0.6"
flake8-bugbear = "^23.3.23"
flake8-builtins = "^2.1.0"
flake8-class-attributes-order = "^0.1.3"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.11.1"
flake8-executable = "^2.1.3"
flake8-expression-complexity = "^0.0.11"
flake8-length = "^0.3.1"
flake8-markdown = "^0.4.0"
flake8-multiline-containers = "^0.0.19"
flake8-noqa = "^1.3.1"
flake8-pytest-style = "^1.7.2"
flake8-quotes = "^3.3.2"
flake8-return = "^1.2.0"
flake8-scream = "^0.1.0"
flake8-secure-coding-standard = "^1.4.0"
flake8-simplify = "^0.19.3"
flake8-string-format = "^0.3.0"
flake8-use-fstring = "^1.4"
flake8-useless-assert = "^0.4.4"
grpc-stubs = "^1.53.0.2"
isort = "^5.12.0"
mypy = "^1.1.1"
mypy-protobuf = "3.4.0"
pylint = "^3.1"
yesqa = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
